Emit map fields
...but not really. I don't want to spend time right now learning how Unreal's `TMap<K, V>` works.

So I emit a comment with the Key and Value type and leave the field as an array of 80 bytes.

Note: the alignment of an array of 80 bytes is 1 byte,
but the alignment of TMap is likely 8 bytes (contains pointers?).
So it's possible that a structure containing a TMap field may not be aligned correctly,
and the field immediately before a TMap may not have suitable padding. Something we'll have to look out for when generating
size and alignment verification tests.